Added counting the correct questions feature and showing the score. Cleaned up old files, deleted activities to try the new features, left only the latest one.
