--- app/src/main/assets/Json.txt
@@ -1,31 +1,0 @@
-
-  {
-    "_id": "58446e0c37b3113b6b024be4",
-    "updatedAt": "2016-12-04T19:56:21.420Z",
-    "createdAt": "2016-12-04T19:27:08.914Z",
-    "question": "The TCP protocol is assigned to which OSI layer?",
-    "__v": 4,
-    "answers": [
-      {
-        "answer": "Transport Layer",
-        "solution": true,
-        "_id": "5844744155b10b3d26489773"
-      },
-      {
-        "answer": "Session Layer",
-        "solution": false,
-        "_id": "5844749955b10b3d26489774"
-      },
-      {
-        "answer": "Network Layer",
-        "solution": false,
-        "_id": "5844749f55b10b3d26489775"
-      },
-      {
-        "answer": "Presentation Layer",
-        "solution": false,
-        "_id": "584474e555b10b3d26489776"
-      }
-    ],
-    "category": "Information technology"
-  }
