+++ app/src/main/assets/twoJsons.txt
@@ -33,7 +33,7 @@
 		"solution": false,
 		"_id": "5844744155b10b3d26489773"
 	}, {
-		"answer": "another answer",
+		"answer": "correct answer",
 		"solution": true,
 		"_id": "5844749955b10b3d26489774"
 	}, {
@@ -57,7 +57,7 @@
    		"solution": false,
    		"_id": "5844744155b10b3d26489773"
    	}, {
-   		"answer": "another answer",
+   		"answer": "correct answer",
    		"solution": true,
    		"_id": "5844749955b10b3d26489774"
    	}, {
@@ -75,14 +75,14 @@
       	"_id": "58446e0c37b3113b6b024be4",
       	"updatedAt": "2016-12-04T19:56:21.420Z",
       	"createdAt": "2016-12-04T19:27:08.914Z",
-      	"question": "Third question",
+      	"question": "Fourth question",
       	"__v": 4,
       	"answers": [{
       		"answer": "some answer",
       		"solution": false,
       		"_id": "5844744155b10b3d26489773"
       	}, {
-      		"answer": "another answer",
+      		"answer": "correct answer",
       		"solution": true,
       		"_id": "5844749955b10b3d26489774"
       	}, {
