--- app/src/main/java/quizme/olgamrost/com/quizme/FinalActivity.java
@@ -1,16 +1,0 @@
-package quizme.olgamrost.com.quizme;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-/**
- * Created by olgamrost on 22/12/2016.
- */
-
-public class FinalActivity extends Activity {
-
-    protected void onCreate(Bundle bundle) {
-        super.onCreate(bundle);
-        setContentView(R.layout.activity_final);
-    }
-}
