+++ app/src/main/java/quizme/olgamrost/com/quizme/Json/Response.java
@@ -1,13 +1,14 @@
-package quizme.olgamrost.com.quizme;
+package quizme.olgamrost.com.quizme.Json;
 
-import java.io.Serializable;
 import java.util.List;
+
+import quizme.olgamrost.com.quizme.Json.AnswersBean;
 
 /**
  * Created by olgamrost on 08/12/2016.
  */
 
-public class Response implements Serializable{
+public class Response {
 
 
     /**
@@ -84,39 +85,4 @@ public class Response implements Serializable{
         this.answers = answers;
     }
 
-//    public  class AnswersBean {
-//        /**
-//         * answer : Transport Layer
-//         * solution : true
-//         * _id : 5844744155b10b3d26489773
-//         */
-//
-//        private String answer;
-//        private boolean solution;
-//        private String _id;
-//
-//        public String getAnswer() {
-//            return answer;
-//        }
-//
-//        public void setAnswer(String answer) {
-//            this.answer = answer;
-//        }
-//
-//        public boolean isSolution() {
-//            return solution;
-//        }
-//
-//        public void setSolution(boolean solution) {
-//            this.solution = solution;
-//        }
-//
-//        public String get_id() {
-//            return _id;
-//        }
-//
-//        public void set_id(String _id) {
-//            this._id = _id;
-//        }
-//    }
 }
