+++ app/src/main/res/layout/activity_show_quiz.xml
@@ -5,14 +5,6 @@
     android:orientation="vertical">
 
     <TextView
-        android:id="@+id/questionNumber"
-        android:text="1"
-        android:layout_marginLeft="20dp"
-        android:layout_marginTop="20dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-    <TextView
         android:id="@+id/question"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
