--- app/src/main/res/layout/showanswer.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:layout_margin="20dp">
-
-    <TextView
-        android:id="@+id/questionOnShowAnswer"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/Question1"
-        android:textSize="30dp"
-        android:layout_margin="20dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/answerTrueOrFalse"
-        android:layout_margin="20dp"
-        android:textSize="20dp"
-        android:layout_gravity="center"/>
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:onClick="showQuestion"
-        android:text="@string/showNextQuestion"
-        android:textSize="20dp"
-        android:layout_gravity="center"/>
-</LinearLayout>
